@@ -18,31 +18,6 @@ export const soapBody = 'soapenv:Body';
 export const soapHeader = 'soapenv:Header';
 export const action = 'executeAnonymous';
 
-export const soapTemplate = `<env:Envelope xmlns:xsd="http://www.w3.org/2001/XMLSchema"
-xmlns:env="http://schemas.xmlsoap.org/soap/envelope/"
-xmlns:cmd="http://soap.sforce.com/2006/08/apex"
-xmlns:apex="http://soap.sforce.com/2006/08/apex">
-    <env:Header>
-        <cmd:SessionHeader>
-            <cmd:sessionId>%s</cmd:sessionId>
-        </cmd:SessionHeader>
-        %s
-    </env:Header>
-    <env:Body>
-        <%s xmlns="http://soap.sforce.com/2006/08/apex">
-            %s
-        </%s>
-    </env:Body>
-</env:Envelope>`;
-
-export const xmlCharMap: { [index: string]: string } = {
-  '<': '&lt;',
-  '>': '&gt;',
-  '&': '&amp;',
-  '"': '&quot;',
-  "'": '&apos;'
-};
-
 export interface SoapResponse {
   [soapEnv]?: {
     [soapHeader]?: { DebuggingInfo: DebuggingInfo };
@@ -52,15 +27,8 @@ export interface SoapResponse {
   };
 }
 
-export interface DebuggingInfo {
+interface DebuggingInfo {
   debugLog: string;
-}
-
-export interface RequestData {
-  method: string;
-  url: string;
-  body: string;
-  headers: {};
 }
 
 export type ExecuteAnonymousResponse = {
