@@ -10,7 +10,5 @@ export { ApexLogGetOptions, LogRecord, LogResult } from './types';
 export {
   DEFAULT_DEBUG_LEVEL_NAME,
   LOG_TYPE,
-  MAX_NUM_LOGS,
-  LOG_TIMER_LENGTH_MINUTES,
-  LISTENER_ABORTED_ERROR_NAME
+  LOG_TIMER_LENGTH_MINUTES
 } from './constants';
