@@ -5,7 +5,7 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-export type StreamingEvent = {
+type StreamingEvent = {
   createdDate: string;
   replayId?: number;
   type: string;
