@@ -5,7 +5,7 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import { Connection } from '@salesforce/core';
-import { JsonCollection } from '@salesforce/ts-types';
+import type { JsonCollection } from '@salesforce/ts-types';
 import { xmlCharMap } from './types';
 import type { HttpRequest } from '@jsforce/jsforce-node';
 
@@ -17,7 +17,5 @@ export async function refreshAuth(
 }
 
 export function escapeXml(data: string): string {
-  return data.replace(/[<>&'"]/g, (char) => {
-    return xmlCharMap[char];
-  });
+  return data.replace(/[<>&'"]/g, (char) => xmlCharMap[char]);
 }
