@@ -7,20 +7,7 @@
 
 import * as fs from 'fs';
 import * as path from 'path';
-import { AnyJson } from '@salesforce/ts-types';
-
-export function ensureDirectoryExists(filePath: string): void {
-  if (fs.existsSync(filePath)) {
-    return;
-  }
-  ensureDirectoryExists(path.dirname(filePath));
-  fs.mkdirSync(filePath);
-}
-
-export function ensureFileExists(filePath: string): void {
-  ensureDirectoryExists(path.dirname(filePath));
-  fs.closeSync(fs.openSync(filePath, 'w'));
-}
+import type { AnyJson } from '@salesforce/ts-types';
 
 /**
  * Method to save a file on disk.
@@ -29,36 +16,9 @@ export function ensureFileExists(filePath: string): void {
  * @param fileContent file contents
  */
 export function createFile(filePath: string, fileContent: AnyJson): void {
-  ensureFileExists(filePath);
+  fs.mkdirSync(path.dirname(filePath), { recursive: true });
+  fs.closeSync(fs.openSync(filePath, 'w'));
 
   const writeStream = fs.createWriteStream(filePath);
   writeStream.write(fileContent);
-}
-
-function streamPromise(stream: fs.WriteStream): Promise<void> {
-  return new Promise((resolve, reject) => {
-    stream.on('end', () => {
-      resolve();
-    });
-    stream.on('error', (error) => {
-      reject(error);
-    });
-  });
-}
-
-/**
- * Method to save multiple files on disk
- * @param fileMap key = filePath, value = file contents
- */
-export async function createFiles(
-  fileMap: { path: string; content: string }[]
-): Promise<void> {
-  const writePromises = fileMap.map((file) => {
-    ensureFileExists(file.path);
-    const writeStream = fs.createWriteStream(file.path);
-    writeStream.write(file.content);
-    return streamPromise(writeStream);
-  });
-
-  await Promise.all(writePromises);
 }
