@@ -11,7 +11,10 @@ import {
   testResults,
   coverageResult,
   coverageFailResult,
-  setupResult
+  setupResult,
+  testResultsWithCategory,
+  coverageResultWithCategory,
+  coverageFailResultWithCategory
 } from './testResults';
 
 describe('Human Reporter Tests', () => {
@@ -150,5 +153,91 @@ describe('Human Reporter Tests', () => {
       'AnimalLocatorTest.testMissingAnimal                      Fail              System.AssertException: Assertion Failed: Should not have found an animal: Expected: FooBar, Actual:'
     );
     expect(result).to.contain('=== Test Summary');
+  });
+
+  // Tests for showCategory functionality
+  it('should show category column when showCategory is true', () => {
+    const result = reporter.format(testResultsWithCategory, false, false, true);
+    expect(result).to.not.be.empty;
+    expect(result).to.contain('=== Test Results');
+    expect(result).to.contain('CATEGORY');
+    expect(result).to.contain('Apex');
+    expect(result).to.contain('Flow');
+    expect(result).to.contain('AccountServiceTest.should_create_account');
+    expect(result).to.contain('AwesomeCalculatorTest.testCallout');
+    expect(result).to.contain('AnimalLocatorTest.testMissingAnimal');
+  });
+
+  it('should not show category column when showCategory is false', () => {
+    const result = reporter.format(
+      testResultsWithCategory,
+      false,
+      false,
+      false
+    );
+    expect(result).to.not.be.empty;
+    expect(result).to.contain('=== Test Results');
+    expect(result).to.not.contain('CATEGORY');
+    // Should still contain the test names but not the category column
+    expect(result).to.contain('AccountServiceTest.should_create_account');
+    expect(result).to.contain('AwesomeCalculatorTest.testCallout');
+    expect(result).to.contain('AnimalLocatorTest.testMissingAnimal');
+  });
+
+  it('should show category column in detailed coverage when showCategory is true', () => {
+    const result = reporter.format(
+      coverageResultWithCategory,
+      true,
+      false,
+      true
+    );
+    expect(result).to.not.be.empty;
+    expect(result).to.contain(
+      '=== Apex Code Coverage for Test Run 7073t000061uwZI'
+    );
+    expect(result).to.contain('CATEGORY');
+    expect(result).to.contain('Apex');
+    expect(result).to.contain('Flow');
+    expect(result).to.contain('AccountService');
+    expect(result).to.contain('CalculatorUtils');
+  });
+
+  it('should not show category column in detailed coverage when showCategory is false', () => {
+    const result = reporter.format(
+      coverageResultWithCategory,
+      true,
+      false,
+      false
+    );
+    expect(result).to.not.be.empty;
+    expect(result).to.contain(
+      '=== Apex Code Coverage for Test Run 7073t000061uwZI'
+    );
+    expect(result).to.not.contain('CATEGORY');
+    // Should still contain the covered class names
+    expect(result).to.contain('AccountService');
+    expect(result).to.contain('CalculatorUtils');
+  });
+
+  it('should show category column with concise mode when showCategory is true', () => {
+    const result = reporter.format(testResultsWithCategory, false, true, true);
+    expect(result).to.not.be.empty;
+    expect(result).to.contain('=== Test Results');
+    expect(result).to.contain('CATEGORY');
+  });
+
+  it('should show category column in detailed coverage with concise mode when showCategory is true', () => {
+    const result = reporter.format(
+      coverageFailResultWithCategory,
+      true,
+      true,
+      true
+    );
+    expect(result).to.not.be.empty;
+    expect(result).to.contain(
+      '=== Apex Code Coverage for Test Run 7073t000061uwZI'
+    );
+    expect(result).to.contain('CATEGORY');
+    expect(result).to.not.contain('=== Apex Code Coverage by Class');
   });
 });
