@@ -12,6 +12,7 @@ import {
   ApexTestResultRecord,
   ApexTestRunResultStatus,
   SyncTestResult,
+  TestCategory,
   TestResult
 } from '../src/tests/types';
 
@@ -178,6 +179,52 @@ export const testResultData: TestResult = {
   setup: []
 };
 
+export const flowTestResultData: TestResult = {
+  // @ts-ignore
+  summary: {
+    failRate: '0%',
+    failing: 0,
+    hostname: 'https://na139.salesforce.com',
+    testsRan: 1,
+    outcome: 'Passed',
+    passRate: '100%',
+    passing: 1,
+    skipRate: '0%',
+    skipped: 0,
+    testStartTime: localStartTime,
+    testSetupTimeInMs: undefined,
+    testExecutionTimeInMs: 0,
+    testTotalTimeInMs: 0,
+    commandTimeInMs: 2000,
+    testRunId,
+    userId: '005xx000000abcDAAU'
+  },
+  tests: [
+    {
+      id: '07Mxx00000F2Xx6UAF',
+      queueItemId: '7092M000000Vt94QAC',
+      stackTrace: '',
+      message: '',
+      asyncApexJobId: '7092M000000Vt94QAC',
+      category: TestCategory.Flow,
+      methodName: 'FlowName_FlowTestName',
+      outcome: ApexTestResultOutcome.Pass,
+      apexLogId: '',
+      apexClass: {
+        id: '7092M000000Vt94QAC',
+        name: 'FlowName',
+        namespacePrefix: '',
+        fullName: 'FlowName'
+      },
+      runTime: 0,
+      testTimestamp: '3',
+      fullName: 'FlowName.FlowName_FlowTestName'
+    }
+  ],
+  setup: [],
+  codecoverage: undefined
+};
+
 export const missingTimeTestData: TestResult = {
   // @ts-ignore
   summary: {
@@ -205,6 +252,7 @@ export const missingTimeTestData: TestResult = {
       stackTrace: null,
       message: null,
       asyncApexJobId: testRunId,
+      category: TestCategory.Apex,
       methodName: 'testLoggerLog',
       outcome: ApexTestResultOutcome.Pass,
       apexLogId: null,
@@ -250,11 +298,12 @@ export const skippedTestData: TestResult = {
       stackTrace: null,
       message: null,
       asyncApexJobId: testRunId,
+      category: TestCategory.Apex,
       methodName: 'testLoggerLog',
       outcome: ApexTestResultOutcome.Skip,
       apexLogId: null,
       apexClass: {
-        id: '01pxx00000O6tXZQAZ',
+        id: '7092M000000Vt94QAC',
         name: 'TestLogger',
         namespacePrefix: 't3st',
         fullName: 't3st.TestLogger'
@@ -297,6 +346,7 @@ export const diagnosticResult: TestResult = {
       stackTrace: 'Class.LIFXControllerTest.makeData: line 6, column 1',
       message: 'System.AssertException: Assertion Failed',
       asyncApexJobId: testRunId,
+      category: TestCategory.Apex,
       methodName: 'testLoggerLog',
       outcome: ApexTestResultOutcome.Fail,
       apexLogId: null,
@@ -334,6 +384,7 @@ export const diagnosticFailure: TestResult = {
       stackTrace: 'Class.LIFXControllerTest.makeData',
       message: 'System.AssertException: Assertion Failed',
       asyncApexJobId: testRunId,
+      category: TestCategory.Apex,
       methodName: 'testLoggerLog',
       outcome: ApexTestResultOutcome.Fail,
       apexLogId: null,
