@@ -31,16 +31,13 @@ import {
   ApexTestResultOutcome,
   ApexTestQueueItem,
   ApexTestRunResultStatus,
-  ApexTestRunResult,
   ApexTestResult,
-  ApexOrgWideCoverage,
-  ApexCodeCoverageAggregate,
-  ApexCodeCoverage,
   ApexTestQueueItemRecord,
   ResultFormat,
-  TestRunIdResult
+  TestRunIdResult,
+  TestCategory
 } from '../../src/tests/types';
-import { AsyncTestRun, StreamingClient } from '../../src/streaming';
+import { StreamingClient } from '../../src/streaming';
 import { fail } from 'assert';
 import { nls } from '../../src/i18n';
 import {
@@ -53,7 +50,8 @@ import {
   testStartTime,
   diagnosticFailure,
   diagnosticResult,
-  skippedTestData
+  skippedTestData,
+  flowTestResultData
 } from '../testData';
 import { join } from 'path';
 import fs from 'node:fs/promises';
@@ -62,31 +60,30 @@ import * as utils from '../../src/tests/utils';
 import { AsyncTests } from '../../src/tests/asyncTests';
 import { QUERY_RECORD_LIMIT } from '../../src/tests/constants';
 import { Writable } from 'node:stream';
+import { Duration } from '@salesforce/kit';
 
 let mockConnection: Connection;
 let sandboxStub: SinonSandbox;
 let toolingRequestStub: SinonStub;
 let retrieveMaxApiVersionStub: SinonStub;
-let singleRecordQueryStub: SinonStub;
+let formatSpy: SinonSpy;
 const testData = new MockTestOrgData();
+let timeStub: SinonStub;
+const pollResponse: ApexTestQueueItem = {
+  done: true,
+  totalSize: 1,
+  records: [
+    {
+      Id: '7092M000000Vt94QAC',
+      Status: ApexTestQueueItemStatus.Completed,
+      ApexClassId: '01p2M00000O6tXZQAZ',
+      TestRunResultId: '05m2M000000TgYuQAK'
+    }
+  ]
+};
 
 describe('Run Apex tests asynchronously', () => {
   const $$ = new TestContext();
-
-  let timeStub: SinonStub;
-  let formatSpy: SinonSpy;
-  const pollResponse: ApexTestQueueItem = {
-    done: true,
-    totalSize: 1,
-    records: [
-      {
-        Id: '7092M000000Vt94QAC',
-        Status: ApexTestQueueItemStatus.Completed,
-        ApexClassId: '01p2M00000O6tXZQAZ',
-        TestRunResultId: '05m2M000000TgYuQAK'
-      }
-    ]
-  };
 
   beforeEach(async () => {
     sandboxStub = createSandbox();
@@ -96,9 +93,9 @@ describe('Run Apex tests asynchronously', () => {
     retrieveMaxApiVersionStub = sandboxStub
       .stub(mockConnection, 'retrieveMaxApiVersion')
       .resolves('61.0');
-    sandboxStub.stub(mockConnection, 'instanceUrl').get(() => {
-      return 'https://na139.salesforce.com';
-    });
+    sandboxStub
+      .stub(mockConnection, 'instanceUrl')
+      .get(() => 'https://na139.salesforce.com');
     sandboxStub.stub(mockConnection, 'getApiVersion').resolves('50.0');
     timeStub = sandboxStub
       .stub(Date.prototype, 'getTime')
@@ -108,10 +105,6 @@ describe('Run Apex tests asynchronously', () => {
     testResultData.summary.orgId = mockConnection.getAuthInfoFields().orgId;
     testResultData.summary.username = mockConnection.getUsername();
     toolingRequestStub = sandboxStub.stub(mockConnection.tooling, 'request');
-    singleRecordQueryStub = sandboxStub.stub(
-      mockConnection,
-      'singleRecordQuery'
-    );
     formatSpy = sandboxStub.spy(diagnosticUtil, 'formatTestErrors');
   });
 
@@ -120,10 +113,6 @@ describe('Run Apex tests asynchronously', () => {
   });
 
   it('should run a successful test', async () => {
-    const asyncResult = {
-      runId: testRunId,
-      queueItem: pollResponse
-    } as AsyncTestRun;
     const requestOptions: AsyncTestConfiguration = {
       classNames: 'TestSample',
       testLevel: TestLevel.RunSpecifiedTests
@@ -137,22 +126,54 @@ describe('Run Apex tests asynchronously', () => {
     };
 
     toolingRequestStub.withArgs(testAsyncRequest).returns(testRunId);
-    sandboxStub
-      .stub(StreamingClient.prototype, 'subscribe')
-      .resolves(asyncResult);
-    sandboxStub.stub(AsyncTests.prototype, 'checkRunStatus').resolves({
-      testsComplete: true,
-      testRunSummary: {} as ApexTestRunResult
+
+    sandboxStub.stub(mockConnection.tooling, 'query').resolves({
+      done: true,
+      totalSize: 1,
+      records: [
+        {
+          Id: '07Mxx00000F2Xx6UAF',
+          AsyncApexJobId: testRunId,
+          Status: ApexTestRunResultStatus.Completed,
+          StartTime: testStartTime,
+          TestTime: 1000,
+          UserId: '005xx000000abcDAAU',
+          ClassesCompleted: 1,
+          ClassesEnqueued: 1,
+          MethodsEnqueued: 1,
+          QueueItemId: '7092M000000Vt94QAC',
+          StackTrace: null,
+          Message: null,
+          MethodName: 'testLoggerLog',
+          Outcome: ApexTestResultOutcome.Pass,
+          ApexLogId: null,
+          ApexClass: {
+            Id: '01pxx00000O6tXZQAZ',
+            Name: 'TestLogger',
+            NamespacePrefix: 't3st',
+            FullName: 't3st__TestLogger'
+          },
+          RunTime: 8,
+          TestTimestamp: '3'
+        }
+      ]
     });
+
     const testSrv = new TestService(mockConnection);
     const mockTestResultData = sandboxStub
       .stub(AsyncTests.prototype, 'formatAsyncResults')
       .resolves(testResultData);
-    sandboxStub.stub(StreamingClient.prototype, 'handshake').resolves();
-    const testResult = await testSrv.runTestAsynchronous(requestOptions);
+
+    const testResult = await testSrv.runTestAsynchronous(
+      requestOptions,
+      false,
+      false,
+      undefined,
+      undefined,
+      Duration.minutes(2)
+    );
     expect(testResult).to.be.a('object');
     expect(mockTestResultData.calledOnce).to.equal(true);
-    expect(mockTestResultData.getCall(0).args[0]).to.equal(asyncResult);
     expect(testResult).to.equal(testResultData);
   });
 
@@ -161,17 +182,157 @@ describe('Run Apex tests asynchronously', () => {
       classNames: 'TestSample',
       testLevel: TestLevel.RunSpecifiedTests
     };
+    const testAsyncRequest = {
+      method: 'POST',
+      url: `${mockConnection.tooling._baseUrl()}/runTestsAsynchronous`,
+      body: JSON.stringify(requestOptions),
+      headers: {
+        'content-type': 'application/json'
+      }
+    };
+    toolingRequestStub.withArgs(testAsyncRequest).returns(testRunId);
 
     sandboxStub
-      .stub(StreamingClient.prototype, 'init')
-      .throwsException('No access token');
+      .stub(mockConnection.tooling, 'query')
+      .rejects(new Error('No access token'));
+
     const testSrv = new TestService(mockConnection);
     try {
-      await testSrv.runTestAsynchronous(requestOptions);
-      fail('Test should have thrown an error');
-    } catch (e) {
-      expect(e.name).to.equal('No access token');
+      await testSrv.runTestAsynchronous(
+        requestOptions,
+        false,
+        false,
+        undefined,
+        undefined,
+        Duration.minutes(2)
+      );
+      assert.fail('Expected an error to be thrown');
+    } catch (error) {
+      assert.equal(error.message, 'No access token');
     }
+  });
+
+  it('should return flow test formatted test results', async () => {
+    flowTestResultData.summary.orgId = mockConnection.getAuthInfoFields().orgId;
+    flowTestResultData.summary.username = mockConnection.getUsername();
+    const asyncTestSrv = new AsyncTests(mockConnection);
+    const mockToolingQuery = sandboxStub.stub(mockConnection.tooling, 'query');
+
+    // First call - test run summary
+    mockToolingQuery.onFirstCall().resolves({
+      done: true,
+      totalSize: 1,
+      records: [
+        {
+          AsyncApexJobId: testRunId,
+          Status: ApexTestRunResultStatus.Completed,
+          StartTime: testStartTime,
+          TestTime: null,
+          UserId: '005xx000000abcDAAU'
+        }
+      ]
+    });
+
+    // Second call - flow test results
+    mockToolingQuery.onSecondCall().resolves({
+      done: true,
+      totalSize: 1,
+      records: [
+        {
+          Id: '07Mxx00000F2Xx6UAF',
+          ApexTestQueueItemId: '7092M000000Vt94QAC',
+          Result: ApexTestResultOutcome.Pass,
+          TestStartDateTime: '3',
+          TestEndDateTime: '5',
+          FlowTest: {
+            DeveloperName: 'FlowName_FlowTestName'
+          },
+          FlowDefinition: {
+            DeveloperName: 'FlowName',
+            NamespacePrefix: ''
+          }
+        }
+      ]
+    });
+
+    const runResult = await asyncTestSrv.checkRunStatus(testRunId);
+    const pollResponseForFlowTest: ApexTestQueueItem = {
+      done: true,
+      totalSize: 1,
+      records: [
+        {
+          Id: '7092M000000Vt94QAC',
+          Status: ApexTestQueueItemStatus.Completed,
+          ApexClassId: null,
+          TestRunResultId: '05m2M000000TgYuQAK'
+        }
+      ]
+    };
+    const getTestResultData = await asyncTestSrv.formatAsyncResults(
+      { queueItem: pollResponseForFlowTest, runId: testRunId },
+      new Date().getTime(),
+      undefined,
+      runResult.testRunSummary
+    );
+
+    let summaryQuery =
+      'SELECT AsyncApexJobId, Status, ClassesCompleted, ClassesEnqueued, ';
+    summaryQuery +=
+      'MethodsEnqueued, StartTime, EndTime, TestTime, TestSetupTime, UserId ';
+    summaryQuery += `FROM ApexTestRunResult WHERE AsyncApexJobId = '${testRunId}'`;
+    expect(mockToolingQuery.getCall(0).args[0]).to.equal(summaryQuery);
+
+    let testResultQuery =
+      'SELECT Id, ApexTestQueueItemId, Result, TestStartDateTime,TestEndDateTime, FlowTest.DeveloperName, ';
+    testResultQuery +=
+      'FlowDefinition.DeveloperName, FlowDefinition.NamespacePrefix ';
+    testResultQuery += `FROM FlowTestResult WHERE ApexTestQueueItemId IN ('${pollResponseForFlowTest.records[0].Id}')`;
+    expect(mockToolingQuery.getCall(1).args[0]).to.equal(testResultQuery);
+    expect(getTestResultData).to.deep.equals(flowTestResultData);
+  });
+
+  it('should return correct summary outcome for single skipped test', async () => {
+    skippedTestData.summary.orgId = mockConnection.getAuthInfoFields().orgId;
+    skippedTestData.summary.username = mockConnection.getUsername();
+    const asyncTestSrv = new AsyncTests(mockConnection);
+    sandboxStub.stub(mockConnection.tooling, 'query').resolves({
+      done: true,
+      totalSize: 1,
+      records: [
+        {
+          Id: '07Mxx00000F2Xx6UAF',
+          AsyncApexJobId: testRunId,
+          Status: ApexTestRunResultStatus.Completed,
+          StartTime: testStartTime,
+          TestTime: null,
+          UserId: '005xx000000abcDAAU',
+          QueueItemId: '7092M000000Vt94QAC',
+          StackTrace: null,
+          Message: null,
+          MethodName: 'testLoggerLog',
+          Outcome: ApexTestResultOutcome.Skip,
+          ApexLogId: null,
+          ApexClass: {
+            Id: '7092M000000Vt94QAC',
+            Name: 'TestLogger',
+            NamespacePrefix: 't3st',
+            FullName: 't3st__TestLogger'
+          },
+          RunTime: null,
+          TestTimestamp: '3'
+        }
+      ]
+    });
+
+    const runResult = await asyncTestSrv.checkRunStatus(testRunId);
+    const testResultData = await asyncTestSrv.formatAsyncResults(
+      { queueItem: pollResponse, runId: testRunId },
+      new Date().getTime(),
+      false,
+      runResult.testRunSummary
+    );
+
+    expect(testResultData).to.deep.equals(skippedTestData);
   });
 
   it('should return formatted test results', async () => {
@@ -179,25 +340,20 @@ describe('Run Apex tests asynchronously', () => {
       mockConnection.getAuthInfoFields().orgId;
     missingTimeTestData.summary.username = mockConnection.getUsername();
     const asyncTestSrv = new AsyncTests(mockConnection);
-    const mockToolingQuery = sandboxStub.stub(mockConnection.tooling, 'query');
-    singleRecordQueryStub.onFirstCall().resolves({
-      AsyncApexJobId: testRunId,
-      Status: ApexTestRunResultStatus.Completed,
-      StartTime: testStartTime,
-      TestTime: null,
-      UserId: '005xx000000abcDAAU'
-    });
-
-    mockToolingQuery.onFirstCall().resolves({
+    sandboxStub.stub(mockConnection.tooling, 'query').resolves({
       done: true,
       totalSize: 1,
       records: [
         {
           Id: '07Mxx00000F2Xx6UAF',
+          AsyncApexJobId: testRunId,
+          Status: ApexTestRunResultStatus.Completed,
+          StartTime: testStartTime,
+          TestTime: null,
+          UserId: '005xx000000abcDAAU',
           QueueItemId: '7092M000000Vt94QAC',
           StackTrace: null,
           Message: null,
-          AsyncApexJobId: testRunId,
           MethodName: 'testLoggerLog',
           Outcome: ApexTestResultOutcome.Pass,
           ApexLogId: null,
@@ -211,7 +367,8 @@ describe('Run Apex tests asynchronously', () => {
           TestTimestamp: '3'
         }
       ]
-    } as ApexTestResult);
+    });
+
     const runResult = await asyncTestSrv.checkRunStatus(testRunId);
     const getTestResultData = await asyncTestSrv.formatAsyncResults(
       { queueItem: pollResponse, runId: testRunId },
@@ -220,34 +377,26 @@ describe('Run Apex tests asynchronously', () => {
       runResult.testRunSummary
     );
 
-    let summaryQuery =
-      'SELECT AsyncApexJobId, Status, ClassesCompleted, ClassesEnqueued, ';
-    summaryQuery +=
-      'MethodsEnqueued, StartTime, EndTime, TestTime, TestSetupTime, UserId ';
-    summaryQuery += `FROM ApexTestRunResult WHERE AsyncApexJobId = '${testRunId}'`;
-    expect(singleRecordQueryStub.getCall(0).args[0]).to.equal(summaryQuery);
-
-    let testResultQuery = 'SELECT Id, QueueItemId, StackTrace, Message, ';
-    testResultQuery +=
-      'RunTime, TestTimestamp, AsyncApexJobId, MethodName, Outcome, ApexLogId, IsTestSetup, ';
-    testResultQuery +=
-      'ApexClass.Id, ApexClass.Name, ApexClass.NamespacePrefix ';
-    testResultQuery += `FROM ApexTestResult WHERE QueueItemId IN ('${pollResponse.records[0].Id}')`;
-    expect(mockToolingQuery.getCall(0).args[0]).to.equal(testResultQuery);
     expect(getTestResultData).to.deep.equals(missingTimeTestData);
   });
 
   it('should report progress when checking test summary for run', async () => {
     const asyncTestSrv = new AsyncTests(mockConnection);
     const mockToolingQuery = sandboxStub.stub(mockConnection.tooling, 'query');
-    singleRecordQueryStub.onFirstCall().resolves({
-      AsyncApexJobId: testRunId,
-      Status: ApexTestRunResultStatus.Completed,
-      StartTime: testStartTime,
-      TestTime: null,
-      UserId: '005xx000000abcDAAU'
-    });
     mockToolingQuery.onFirstCall().resolves({
+      done: true,
+      totalSize: 1,
+      records: [
+        {
+          AsyncApexJobId: testRunId,
+          Status: ApexTestRunResultStatus.Completed,
+          StartTime: testStartTime,
+          TestTime: null,
+          UserId: '005xx000000abcDAAU'
+        }
+      ]
+    });
+    mockToolingQuery.onSecondCall().resolves({
       done: true,
       totalSize: 1,
       records: [
@@ -286,83 +435,25 @@ describe('Run Apex tests asynchronously', () => {
     });
   });
 
-  it('should return correct summary outcome for single skipped test', async () => {
-    skippedTestData.summary.orgId = mockConnection.getAuthInfoFields().orgId;
-    skippedTestData.summary.username = mockConnection.getUsername();
-    const asyncTestSrv = new AsyncTests(mockConnection);
-    const mockToolingQuery = sandboxStub.stub(mockConnection.tooling, 'query');
-    singleRecordQueryStub.onFirstCall().resolves({
-      AsyncApexJobId: testRunId,
-      Status: ApexTestRunResultStatus.Completed,
-      StartTime: testStartTime,
-      TestTime: null,
-      UserId: '005xx000000abcDAAU'
-    });
-
-    mockToolingQuery.onFirstCall().resolves({
-      done: true,
-      totalSize: 1,
-      records: [
-        {
-          Id: '07Mxx00000F2Xx6UAF',
-          QueueItemId: '7092M000000Vt94QAC',
-          StackTrace: null,
-          Message: null,
-          AsyncApexJobId: testRunId,
-          MethodName: 'testLoggerLog',
-          Outcome: ApexTestResultOutcome.Skip,
-          ApexLogId: null,
-          ApexClass: {
-            Id: '01pxx00000O6tXZQAZ',
-            Name: 'TestLogger',
-            NamespacePrefix: 't3st',
-            FullName: 't3st__TestLogger'
-          },
-          RunTime: null,
-          TestTimestamp: '3'
-        }
-      ]
-    } as ApexTestResult);
-
-    const runResult = await asyncTestSrv.checkRunStatus(testRunId);
-    const getTestResultData = await asyncTestSrv.formatAsyncResults(
-      { queueItem: pollResponse, runId: testRunId },
-      new Date().getTime(),
-      false,
-      runResult.testRunSummary
-    );
-
-    let summaryQuery =
-      'SELECT AsyncApexJobId, Status, ClassesCompleted, ClassesEnqueued, ';
-    summaryQuery +=
-      'MethodsEnqueued, StartTime, EndTime, TestTime, TestSetupTime, UserId ';
-    summaryQuery += `FROM ApexTestRunResult WHERE AsyncApexJobId = '${testRunId}'`;
-    expect(singleRecordQueryStub.getCall(0).args[0]).to.equal(summaryQuery);
-
-    let testResultQuery = 'SELECT Id, QueueItemId, StackTrace, Message, ';
-    testResultQuery +=
-      'RunTime, TestTimestamp, AsyncApexJobId, MethodName, Outcome, ApexLogId, IsTestSetup, ';
-    testResultQuery +=
-      'ApexClass.Id, ApexClass.Name, ApexClass.NamespacePrefix ';
-    testResultQuery += `FROM ApexTestResult WHERE QueueItemId IN ('${pollResponse.records[0].Id}')`;
-    expect(mockToolingQuery.getCall(0).args[0]).to.equal(testResultQuery);
-    expect(getTestResultData).to.deep.equals(skippedTestData);
-  });
-
   it('should return formatted test results with diagnostics', async () => {
     diagnosticResult.summary.orgId = mockConnection.getAuthInfoFields().orgId;
     diagnosticResult.summary.username = mockConnection.getUsername();
     const asyncTestSrv = new AsyncTests(mockConnection);
     const mockToolingQuery = sandboxStub.stub(mockConnection.tooling, 'query');
-    singleRecordQueryStub.onFirstCall().resolves({
-      AsyncApexJobId: testRunId,
-      Status: ApexTestRunResultStatus.Completed,
-      StartTime: testStartTime,
-      TestTime: null,
-      UserId: '005xx000000abcDAAU'
-    });
-
     mockToolingQuery.onFirstCall().resolves({
+      done: true,
+      totalSize: 1,
+      records: [
+        {
+          AsyncApexJobId: testRunId,
+          Status: ApexTestRunResultStatus.Completed,
+          StartTime: testStartTime,
+          TestTime: null,
+          UserId: '005xx000000abcDAAU'
+        }
+      ]
+    });
+    mockToolingQuery.onSecondCall().resolves({
       done: true,
       totalSize: 1,
       records: [
@@ -406,15 +497,21 @@ describe('Run Apex tests asynchronously', () => {
     diagnosticFailure.tests[0].stackTrace = undefined;
     const asyncTestSrv = new AsyncTests(mockConnection);
     const mockToolingQuery = sandboxStub.stub(mockConnection.tooling, 'query');
-    singleRecordQueryStub.onFirstCall().resolves({
-      AsyncApexJobId: testRunId,
-      Status: ApexTestRunResultStatus.Completed,
-      StartTime: testStartTime,
-      TestTime: null,
-      UserId: '005xx000000abcDAAU'
+    mockToolingQuery.onFirstCall().resolves({
+      done: true,
+      totalSize: 1,
+      records: [
+        {
+          AsyncApexJobId: testRunId,
+          Status: ApexTestRunResultStatus.Completed,
+          StartTime: testStartTime,
+          TestTime: null,
+          UserId: '005xx000000abcDAAU'
+        }
+      ]
     });
 
-    mockToolingQuery.onFirstCall().resolves({
+    mockToolingQuery.onSecondCall().resolves({
       done: true,
       totalSize: 1,
       records: [
@@ -452,7 +549,12 @@ describe('Run Apex tests asynchronously', () => {
 
   it('should return an error if no test results are found', async () => {
     const asyncTestSrv = new AsyncTests(mockConnection);
-    singleRecordQueryStub.onFirstCall().throwsException('No records found');
+    const mockToolingQuery = sandboxStub.stub(mockConnection.tooling, 'query');
+    mockToolingQuery.onFirstCall().resolves({
+      done: true,
+      totalSize: 0,
+      records: []
+    });
 
     try {
       const runResult = await asyncTestSrv.checkRunStatus(testRunId);
@@ -473,7 +575,12 @@ describe('Run Apex tests asynchronously', () => {
   it('should return an error if invalid test run id was provided', async () => {
     const invalidId = '000000xxxxx';
     const asyncTestSrv = new AsyncTests(mockConnection);
-    singleRecordQueryStub.onFirstCall().resolves(undefined);
+    const mockToolingQuery = sandboxStub.stub(mockConnection.tooling, 'query');
+    mockToolingQuery.onFirstCall().resolves({
+      done: true,
+      totalSize: 0,
+      records: []
+    });
 
     try {
       await asyncTestSrv.checkRunStatus(invalidId);
@@ -482,14 +589,19 @@ describe('Run Apex tests asynchronously', () => {
       expect(e.message).to.equal(
         nls.localize('invalidTestRunIdErr', invalidId)
       );
-      expect(singleRecordQueryStub.notCalled).to.be.true;
+      expect(mockToolingQuery.notCalled).to.be.true;
     }
   });
 
   it('should return an error if invalid test run id prefix was provided', async () => {
     const invalidId = '708000000xxxxxx';
     const asyncTestSrv = new AsyncTests(mockConnection);
-    singleRecordQueryStub.onFirstCall().resolves(undefined);
+    const mockToolingQuery = sandboxStub.stub(mockConnection.tooling, 'query');
+    mockToolingQuery.onFirstCall().resolves({
+      done: true,
+      totalSize: 0,
+      records: []
+    });
 
     try {
       await asyncTestSrv.checkRunStatus(invalidId);
@@ -498,43 +610,71 @@ describe('Run Apex tests asynchronously', () => {
       expect(e.message).to.equal(
         nls.localize('invalidTestRunIdErr', invalidId)
       );
-      expect(singleRecordQueryStub.notCalled).to.be.true;
+      expect(mockToolingQuery.notCalled).to.be.true;
     }
   });
 
   it('should return formatted test results with code coverage', async () => {
     const asyncTestSrv = new AsyncTests(mockConnection);
-    const mockToolingAutoQuery = sandboxStub.stub(
-      mockConnection.tooling,
-      'query'
-    );
+    const mockToolingQuery = sandboxStub.stub(mockConnection.tooling, 'query');
 
-    singleRecordQueryStub.onFirstCall().resolves({
-      AsyncApexJobId: testRunId,
-      Status: ApexTestRunResultStatus.Completed,
-      StartTime: '2020-07-12T02:54:47.000+0000',
-      TestTime: 1765,
-      UserId: '005xx000000abcDAAU'
+    // First call - test run summary
+    mockToolingQuery.onFirstCall().resolves({
+      done: true,
+      totalSize: 1,
+      records: [
+        {
+          AsyncApexJobId: testRunId,
+          Status: ApexTestRunResultStatus.Completed,
+          StartTime: testStartTime,
+          TestTime: null,
+          UserId: '005xx000000abcDAAU'
+        }
+      ]
     });
-    mockToolingAutoQuery.onCall(0).resolves({
+
+    // Second call - test results
+    mockToolingQuery.onSecondCall().resolves({
       done: true,
       totalSize: 6,
-      records: mixedTestResults
-    } as ApexTestResult);
+      records: mixedTestResults.map((result) => ({
+        ...result,
+        ApexClass: {
+          Id: result.ApexClass?.Id || '01pxx00000O6tXZQAZ',
+          Name: result.ApexClass?.Name || 'TestLogger',
+          FullName: result.ApexClass?.FullName || 't3st__TestLogger'
+        }
+      }))
+    });
 
-    mockToolingAutoQuery.onCall(1).resolves({
+    // Third call - per-class code coverage
+    mockToolingQuery.onCall(2).resolves({
       done: true,
       totalSize: 3,
-      records: mixedPerClassCodeCoverage
-    } as ApexCodeCoverage);
+      records: mixedPerClassCodeCoverage.map((coverage) => ({
+        ...coverage,
+        ApexClassOrTrigger: {
+          Id: coverage.ApexClassOrTrigger?.Id || '01pxx00000O6tXZQAZ',
+          Name: coverage.ApexClassOrTrigger?.Name || 'TestLogger'
+        }
+      }))
+    });
 
-    mockToolingAutoQuery.onCall(2).resolves({
+    // Fourth call - code coverage aggregate
+    mockToolingQuery.onCall(3).resolves({
       done: true,
       totalSize: 3,
-      records: codeCoverageQueryResult
-    } as ApexCodeCoverageAggregate);
+      records: codeCoverageQueryResult.map((result) => ({
+        ...result,
+        ApexClassOrTrigger: {
+          Id: result.ApexClassOrTrigger?.Id || '01pxx00000O6tXZQAZ',
+          Name: result.ApexClassOrTrigger?.Name || 'TestLogger'
+        }
+      }))
+    });
 
-    mockToolingAutoQuery.onCall(3).resolves({
+    // Fifth call - org-wide coverage
+    mockToolingQuery.onCall(4).resolves({
       done: true,
       totalSize: 1,
       records: [
@@ -542,7 +682,7 @@ describe('Run Apex tests asynchronously', () => {
           PercentCovered: '57'
         }
       ]
-    } as ApexOrgWideCoverage);
+    });
 
     const runResult = await asyncTestSrv.checkRunStatus(testRunId);
     const getTestResultData = await asyncTestSrv.formatAsyncResults(
@@ -572,42 +712,45 @@ describe('Run Apex tests asynchronously', () => {
 
   it('should report progress for aggregating code coverage', async () => {
     const asyncTestSrv = new AsyncTests(mockConnection);
-    const mockToolingQuery = sandboxStub.stub(mockConnection.tooling, 'query');
-    singleRecordQueryStub.onCall(0).resolves({
-      AsyncApexJobId: testRunId,
-      Status: ApexTestRunResultStatus.Completed,
-      StartTime: '2020-07-12T02:54:47.000+0000',
-      TestTime: 1765,
-      UserId: '005xx000000abcDAAU'
-    });
-
-    mockToolingQuery.onCall(0).resolves({
-      done: true,
-      totalSize: 6,
-      records: mixedTestResults
-    } as ApexTestResult);
-
-    mockToolingQuery.onCall(1).resolves({
-      done: true,
-      totalSize: 3,
-      records: mixedPerClassCodeCoverage
-    } as ApexCodeCoverage);
-
-    mockToolingQuery.onCall(2).resolves({
-      done: true,
-      totalSize: 3,
-      records: codeCoverageQueryResult
-    } as ApexCodeCoverageAggregate);
-
-    mockToolingQuery.onCall(3).resolves({
+    sandboxStub.stub(mockConnection.tooling, 'query').resolves({
       done: true,
       totalSize: 1,
       records: [
         {
-          PercentCovered: '57'
+          Id: '7092M000000Vt94QAC',
+          AsyncApexJobId: testRunId,
+          Status: ApexTestRunResultStatus.Completed,
+          StartTime: testStartTime,
+          TestTime: null,
+          UserId: '005xx000000abcDAAU',
+          QueueItemId: '7092M000000Vt94QAC',
+          StackTrace: null,
+          Message: null,
+          MethodName: 'testLoggerLog',
+          Outcome: ApexTestResultOutcome.Pass,
+          ApexLogId: null,
+          ApexClass: {
+            Id: '01pxx00000O6tXZQAZ',
+            Name: 'TestLogger',
+            NamespacePrefix: 't3st',
+            FullName: 't3st__TestLogger'
+          },
+          ApexClassOrTrigger: {
+            Id: '01pxx00000O6tXZQAZ',
+            Name: 'TestLogger'
+          },
+          NumLinesCovered: 10,
+          NumLinesUncovered: 5,
+          PercentCovered: '57',
+          RunTime: 8,
+          TestTimestamp: '3',
+          Coverage: {
+            coveredLines: [1, 2, 3],
+            uncoveredLines: [4, 5]
+          }
         }
       ]
-    } as ApexOrgWideCoverage);
+    });
 
     const reportStub = sandboxStub.stub();
     const progressReporter: Progress<ApexTestProgressValue> = {
@@ -808,6 +951,15 @@ describe('Run Apex tests asynchronously', () => {
         totalSize: 1,
         records: [
           {
+            ApexClassId: 'xxxx'
+          }
+        ]
+      });
+      mockToolingQuery.onSecondCall().resolves({
+        done: true,
+        totalSize: 1,
+        records: [
+          {
             Id: '07Mxx00000F2Xx6UAF',
             QueueItemId: id,
             StackTrace: null,
@@ -912,64 +1064,139 @@ describe('Run Apex tests asynchronously', () => {
         }
       };
       toolingRequestStub.withArgs(testAsyncRequest).returns(testRunId);
-      const actionf: () => Promise<{ runId: string }> = () => {
-        return Promise.resolve({ runId: testRunId });
-      };
 
-      sandboxStub
-        .stub(StreamingClient.prototype, 'subscribe')
-        .callsFake(function () {
-          // eslint-disable-next-line
-          const that = this;
-          return new Promise(function () {
-            actionf().then(function (id) {
-              that.subscribedTestRunId = id;
-              that.subscribedTestRunIdDeferred.resolve(id);
-            });
-          });
-        });
-      const diconnectStub = sandboxStub.stub(
-        StreamingClient.prototype,
-        'disconnect'
+      // Mock the tooling query responses for polling
+      const mockToolingQuery = sandboxStub.stub(
+        mockConnection.tooling,
+        'query'
       );
-      sandboxStub.stub(StreamingClient.prototype, 'handshake').resolves();
+
+      // First call - initial test run status
+      mockToolingQuery.onFirstCall().resolves({
+        done: true,
+        totalSize: 1,
+        records: [
+          {
+            AsyncApexJobId: testRunId,
+            Status: ApexTestRunResultStatus.Queued,
+            StartTime: testStartTime,
+            TestTime: null,
+            UserId: '005xx000000abcDAAU'
+          }
+        ]
+      });
+
+      // Second call - test queue items
+      mockToolingQuery.onSecondCall().resolves({
+        done: true,
+        totalSize: 1,
+        records: [
+          {
+            Id: '7092M000000Vt94QAC',
+            Status: ApexTestQueueItemStatus.Processing,
+            ApexClassId: '01p2M00000O6tXZQAZ',
+            TestRunResultId: '05m2M000000TgYuQAK'
+          }
+        ]
+      });
+
+      // Third call - test run summary after abort
+      mockToolingQuery.onThirdCall().resolves({
+        done: true,
+        totalSize: 1,
+        records: [
+          {
+            AsyncApexJobId: testRunId,
+            Status: ApexTestRunResultStatus.Aborted,
+            StartTime: testStartTime,
+            TestTime: null,
+            UserId: '005xx000000abcDAAU'
+          }
+        ]
+      });
+
       const abortTestRunStub = sandboxStub
         .stub(AsyncTests.prototype, 'abortTestRun')
         .resolves();
 
       const cancellationTokenSource = new CancellationTokenSource();
       const testSrv = new TestService(mockConnection);
-      testSrv.runTestAsynchronous(
+      const testPromise = testSrv.runTestAsynchronous(
         requestOptions,
         false,
         undefined,
         undefined,
-        cancellationTokenSource.token
+        cancellationTokenSource.token,
+        Duration.minutes(2)
       );
 
-      return new Promise((resolve) => {
-        // wait for task queue
-        setTimeout(async () => {
-          await cancellationTokenSource.asyncCancel();
-          sinonAssert.calledOnce(abortTestRunStub);
-          sinonAssert.calledOnce(diconnectStub);
-          resolve();
-        }, 100);
-      });
+      // Wait for the test to start
+      await new Promise((resolve) => setTimeout(resolve, 100));
+
+      // Cancel the test
+      await cancellationTokenSource.asyncCancel();
+
+      // Wait for the test to complete
+      await testPromise;
+
+      sinonAssert.calledOnce(abortTestRunStub);
     });
   });
 
   describe('Format Test Errors', async () => {
     it('should format test error when running asynchronous tests', async () => {
+      const requestOptions: AsyncTestConfiguration = {
+        classNames: 'TestSample',
+        testLevel: TestLevel.RunSpecifiedTests
+      };
+      const testAsyncRequest = {
+        method: 'POST',
+        url: `${mockConnection.tooling._baseUrl()}/runTestsAsynchronous`,
+        body: JSON.stringify(requestOptions),
+        headers: {
+          'content-type': 'application/json'
+        }
+      };
+      toolingRequestStub.withArgs(testAsyncRequest).returns(testRunId);
+
+      // Mock the tooling query responses
+      const mockToolingQuery = sandboxStub.stub(
+        mockConnection.tooling,
+        'query'
+      );
+      mockToolingQuery.onFirstCall().resolves({
+        done: true,
+        totalSize: 1,
+        records: [
+          {
+            AsyncApexJobId: testRunId,
+            Status: ApexTestRunResultStatus.Queued,
+            StartTime: testStartTime,
+            TestTime: null,
+            UserId: '005xx000000abcDAAU'
+          }
+        ]
+      });
+      mockToolingQuery
+        .onSecondCall()
+        .rejects(new Error("sObject type 'ApexClass' is not supported."));
+
       const testSrv = new TestService(mockConnection);
       const errMsg = `sObject type 'ApexClass' is not supported.`;
       sandboxStub
         .stub(StreamingClient.prototype, 'handshake')
         .throws(new Error(errMsg));
       try {
-        await testSrv.runTestAsynchronous({
-          testLevel: TestLevel.RunLocalTests
-        });
+        await testSrv.runTestAsynchronous(
+          {
+            testLevel: TestLevel.RunLocalTests
+          },
+          false,
+          false,
+          undefined,
+          undefined,
+          Duration.minutes(2)
+        );
         fail('Should have failed');
       } catch (e) {
         expect(formatSpy.calledOnce).to.be.true;
@@ -1001,23 +1228,38 @@ describe('Run Apex tests asynchronously', () => {
   describe('Report Test Run Status', async () => {
     it('should subscribe to test run for run still in progress', async () => {
       const asyncTestSrv = new AsyncTests(mockConnection);
-      sandboxStub.stub(mockConnection.tooling, 'query');
-      singleRecordQueryStub
+      const mockToolingQueryProgress = sandboxStub.stub(
+        mockConnection.tooling,
+        'query'
+      );
+      mockToolingQueryProgress
         .onFirstCall()
         .resolves({
-          AsyncApexJobId: testRunId,
-          Status: ApexTestRunResultStatus.Queued,
-          StartTime: testStartTime,
-          TestTime: null,
-          UserId: '005xx000000abcDAAU'
+          done: true,
+          totalSize: 1,
+          records: [
+            {
+              AsyncApexJobId: testRunId,
+              Status: ApexTestRunResultStatus.Queued,
+              StartTime: testStartTime,
+              TestTime: null,
+              UserId: '005xx000000abcDAAU'
+            }
+          ]
         })
         .onSecondCall()
         .resolves({
-          AsyncApexJobId: testRunId,
-          Status: ApexTestRunResultStatus.Completed,
-          StartTime: testStartTime,
-          TestTime: null,
-          UserId: '005xx000000abcDAAU'
+          done: true,
+          totalSize: 1,
+          records: [
+            {
+              AsyncApexJobId: testRunId,
+              Status: ApexTestRunResultStatus.Completed,
+              StartTime: testStartTime,
+              TestTime: null,
+              UserId: '005xx000000abcDAAU'
+            }
+          ]
         });
       const formatResultsStub = sandboxStub.stub(
         asyncTestSrv,
@@ -1049,7 +1291,7 @@ describe('Run Apex tests asynchronously', () => {
 
       await asyncTestSrv.reportAsyncResults(testRunId);
 
-      expect(singleRecordQueryStub.calledTwice).to.be.true;
+      expect(mockToolingQueryProgress.calledTwice).to.be.true;
       expect(formatResultsStub.calledOnce).to.be.true;
       expect(subscribeStub.calledOnce).to.be.true;
       expect(handlerStub.notCalled).to.be.true;
@@ -1057,13 +1299,22 @@ describe('Run Apex tests asynchronously', () => {
 
     it('should query for test run results if run is complete', async () => {
       const asyncTestSrv = new AsyncTests(mockConnection);
-      sandboxStub.stub(mockConnection.tooling, 'query');
-      singleRecordQueryStub.onFirstCall().resolves({
-        AsyncApexJobId: testRunId,
-        Status: ApexTestRunResultStatus.Completed,
-        StartTime: testStartTime,
-        TestTime: null,
-        UserId: '005xx000000abcDAAU'
+      const mockToolingQueryComplete = sandboxStub.stub(
+        mockConnection.tooling,
+        'query'
+      );
+      mockToolingQueryComplete.onFirstCall().resolves({
+        done: true,
+        totalSize: 1,
+        records: [
+          {
+            AsyncApexJobId: testRunId,
+            Status: ApexTestRunResultStatus.Completed,
+            StartTime: testStartTime,
+            TestTime: null,
+            UserId: '005xx000000abcDAAU'
+          }
+        ]
       });
       const formatResultsStub = sandboxStub.stub(
         asyncTestSrv,
@@ -1095,7 +1346,7 @@ describe('Run Apex tests asynchronously', () => {
 
       await asyncTestSrv.reportAsyncResults(testRunId);
 
-      expect(singleRecordQueryStub.calledOnce).to.be.true;
+      expect(mockToolingQueryComplete.calledOnce).to.be.true;
       expect(formatResultsStub.calledOnce).to.be.true;
       expect(subscribeStub.notCalled).to.be.true;
       expect(handlerStub.calledOnce).to.be.true;
@@ -1103,13 +1354,22 @@ describe('Run Apex tests asynchronously', () => {
 
     it('should format results with retrieved test run summary', async () => {
       const asyncTestSrv = new AsyncTests(mockConnection);
-      sandboxStub.stub(mockConnection.tooling, 'query');
-      singleRecordQueryStub.onFirstCall().resolves({
-        AsyncApexJobId: testRunId,
-        Status: ApexTestRunResultStatus.Completed,
-        StartTime: testStartTime,
-        TestTime: null,
-        UserId: '005xx000000abcDAAU'
+      const mockToolingQueryFormat = sandboxStub.stub(
+        mockConnection.tooling,
+        'query'
+      );
+      mockToolingQueryFormat.onFirstCall().resolves({
+        done: true,
+        totalSize: 1,
+        records: [
+          {
+            AsyncApexJobId: testRunId,
+            Status: ApexTestRunResultStatus.Completed,
+            StartTime: testStartTime,
+            TestTime: null,
+            UserId: '005xx000000abcDAAU'
+          }
+        ]
       });
       const formatResultsStub = sandboxStub.stub(
         asyncTestSrv,
@@ -1147,38 +1407,54 @@ describe('Run Apex tests asynchronously', () => {
   describe('Supports Test Setup Feature', async () => {
     it(`should verify org's api version supports test setup feature`, async () => {
       const asyncTests = new AsyncTests(mockConnection);
+      const mockToolingQuery = sandboxStub.stub(
+        mockConnection.tooling,
+        'query'
+      );
 
       const fieldExists = await asyncTests.supportsTestSetupFeature();
       expect(fieldExists).to.be.true;
       expect(retrieveMaxApiVersionStub.calledOnce).to.be.true;
 
       const mockQueryResult = {
-        Status: 'Completed',
-        ClassesCompleted: 5,
-        ClassesEnqueued: 5,
-        MethodsEnqueued: 5,
-        StartTime: '2021-01-01T00:00:00Z',
-        EndTime: '2021-01-01T00:10:00Z',
-        TestTime: 600000,
-        UserId: 'someUserId'
+        done: true,
+        totalSize: 1,
+        records: [
+          {
+            Status: 'Completed',
+            ClassesCompleted: 5,
+            ClassesEnqueued: 5,
+            MethodsEnqueued: 5,
+            StartTime: '2021-01-01T00:00:00Z',
+            EndTime: '2021-01-01T00:10:00Z',
+            TestTime: 600000,
+            UserId: 'someUserId'
+          }
+        ]
       };
 
-      singleRecordQueryStub.resolves(mockQueryResult);
+      mockToolingQuery.resolves(mockQueryResult);
 
       const runStatusResult = await asyncTests.checkRunStatus(testRunId);
 
       expect(runStatusResult.testsComplete).to.be.true;
-      expect(runStatusResult.testRunSummary).to.deep.equal(mockQueryResult);
-      expect(singleRecordQueryStub.calledOnce).to.be.true;
+      expect(runStatusResult.testRunSummary).to.deep.equal(
+        mockQueryResult.records[0]
+      );
+      expect(mockToolingQuery.calledOnce).to.be.true;
 
       const expectedQueryWithTestSetupTime = `SELECT AsyncApexJobId, Status, ClassesCompleted, ClassesEnqueued, MethodsEnqueued, StartTime, EndTime, TestTime, TestSetupTime, UserId FROM ApexTestRunResult WHERE AsyncApexJobId = '${testRunId}'`;
-      expect(singleRecordQueryStub.getCall(0).args[0]).to.equal(
+      expect(mockToolingQuery.getCall(0).args[0]).to.equal(
         expectedQueryWithTestSetupTime
       );
     });
     it('should handle absence of TestSetupTime field and modify query accordingly', async () => {
       retrieveMaxApiVersionStub.resolves('60.0');
       const asyncTests = new AsyncTests(mockConnection);
+      const mockToolingQuery = sandboxStub.stub(
+        mockConnection.tooling,
+        'query'
+      );
 
       const fieldExists = await asyncTests.supportsTestSetupFeature();
 
@@ -1186,28 +1462,347 @@ describe('Run Apex tests asynchronously', () => {
       expect(retrieveMaxApiVersionStub.calledOnce).to.be.true;
 
       const mockQueryResult = {
-        Status: 'Completed',
-        ClassesCompleted: 5,
-        ClassesEnqueued: 5,
-        MethodsEnqueued: 5,
-        StartTime: '2021-01-01T00:00:00Z',
-        EndTime: '2021-01-01T00:10:00Z',
-        TestTime: 600000,
-        UserId: 'someUserId'
+        done: true,
+        totalSize: 1,
+        records: [
+          {
+            Status: 'Completed',
+            ClassesCompleted: 5,
+            ClassesEnqueued: 5,
+            MethodsEnqueued: 5,
+            StartTime: '2021-01-01T00:00:00Z',
+            EndTime: '2021-01-01T00:10:00Z',
+            TestTime: 600000,
+            UserId: 'someUserId'
+          }
+        ]
       };
 
-      singleRecordQueryStub.resolves(mockQueryResult);
+      mockToolingQuery.resolves(mockQueryResult);
 
       const runStatusResult = await asyncTests.checkRunStatus(testRunId);
 
       expect(runStatusResult.testsComplete).to.be.true;
-      expect(runStatusResult.testRunSummary).to.deep.equal(mockQueryResult);
-      expect(singleRecordQueryStub.calledOnce).to.be.true;
+      expect(runStatusResult.testRunSummary).to.deep.equal(
+        mockQueryResult.records[0]
+      );
+      expect(mockToolingQuery.calledOnce).to.be.true;
 
       const expectedQueryWithoutTestSetupTime = `SELECT AsyncApexJobId, Status, ClassesCompleted, ClassesEnqueued, MethodsEnqueued, StartTime, EndTime, TestTime, UserId FROM ApexTestRunResult WHERE AsyncApexJobId = '${testRunId}'`;
-      expect(singleRecordQueryStub.getCall(0).args[0]).to.equal(
+      expect(mockToolingQuery.getCall(0).args[0]).to.equal(
         expectedQueryWithoutTestSetupTime
       );
+    });
+  });
+
+  describe('Polling Client Timeout', () => {
+    it('should return test run ID when polling client times out', async () => {
+      const requestOptions: AsyncTestConfiguration = {
+        classNames: 'TestSample',
+        testLevel: TestLevel.RunSpecifiedTests
+      };
+      const testAsyncRequest = {
+        method: 'POST',
+        url: `${mockConnection.tooling._baseUrl()}/runTestsAsynchronous`,
+        body: JSON.stringify(requestOptions),
+        headers: {
+          'content-type': 'application/json'
+        }
+      };
+      toolingRequestStub.withArgs(testAsyncRequest).returns(testRunId);
+
+      // Mock the tooling query to return a status that will keep polling
+      const mockToolingQuery = sandboxStub.stub(
+        mockConnection.tooling,
+        'query'
+      );
+      mockToolingQuery.onFirstCall().resolves({
+        done: true,
+        totalSize: 1,
+        records: [
+          {
+            AsyncApexJobId: testRunId,
+            Status: ApexTestRunResultStatus.Queued,
+            StartTime: testStartTime,
+            TestTime: null,
+            UserId: '005xx000000abcDAAU'
+          }
+        ]
+      });
+      mockToolingQuery.onSecondCall().resolves({
+        done: true,
+        totalSize: 1,
+        records: [
+          {
+            Id: '7092M000000Vt94QAC',
+            Status: ApexTestQueueItemStatus.Processing,
+            ApexClassId: '01p2M00000O6tXZQAZ',
+            TestRunResultId: '05m2M000000TgYuQAK'
+          }
+        ]
+      });
+
+      // Mock the logger to capture the info message
+      const loggerStub = sandboxStub.stub(Logger.prototype, 'info');
+      const debugStub = sandboxStub.stub(Logger.prototype, 'debug');
+
+      const testSrv = new TestService(mockConnection);
+      const result = await testSrv.runTestAsynchronous(
+        requestOptions,
+        false,
+        false,
+        undefined,
+        undefined,
+        Duration.milliseconds(10) // Very short timeout to trigger timeout quickly
+      );
+
+      // Verify that the result contains the test run ID
+      expect(result).to.have.property('testRunId');
+      expect((result as TestRunIdResult).testRunId).to.equal(testRunId);
+
+      // Verify that the debug message was logged
+      sinonAssert.calledWith(
+        debugStub,
+        sinon.match.string.and(sinon.match(testRunId))
+      );
+
+      // Verify that the info message with the command was logged
+      const username = mockConnection.getUsername();
+      sinonAssert.calledWith(
+        loggerStub,
+        nls.localize('runTestReportCommand', [testRunId, username])
+      );
+    });
+
+    it('should handle PollingClientTimeout error gracefully', async () => {
+      const requestOptions: AsyncTestConfiguration = {
+        classNames: 'TestSample',
+        testLevel: TestLevel.RunSpecifiedTests
+      };
+      const testAsyncRequest = {
+        method: 'POST',
+        url: `${mockConnection.tooling._baseUrl()}/runTestsAsynchronous`,
+        body: JSON.stringify(requestOptions),
+        headers: {
+          'content-type': 'application/json'
+        }
+      };
+      toolingRequestStub.withArgs(testAsyncRequest).returns(testRunId);
+
+      // Mock the tooling query to return a status that will keep polling
+      const mockToolingQuery = sandboxStub.stub(
+        mockConnection.tooling,
+        'query'
+      );
+      mockToolingQuery.onFirstCall().resolves({
+        done: true,
+        totalSize: 1,
+        records: [
+          {
+            AsyncApexJobId: testRunId,
+            Status: ApexTestRunResultStatus.Queued,
+            StartTime: testStartTime,
+            TestTime: null,
+            UserId: '005xx000000abcDAAU'
+          }
+        ]
+      });
+      mockToolingQuery.onSecondCall().resolves({
+        done: true,
+        totalSize: 1,
+        records: [
+          {
+            Id: '7092M000000Vt94QAC',
+            Status: ApexTestQueueItemStatus.Processing,
+            ApexClassId: '01p2M00000O6tXZQAZ',
+            TestRunResultId: '05m2M000000TgYuQAK'
+          }
+        ]
+      });
+
+      // Mock the logger to capture the info message
+      const loggerStub = sandboxStub.stub(Logger.prototype, 'info');
+      const debugStub = sandboxStub.stub(Logger.prototype, 'debug');
+
+      const testSrv = new TestService(mockConnection);
+
+      // This should not throw an error, even though the polling client times out
+      const result = await testSrv.runTestAsynchronous(
+        requestOptions,
+        false,
+        false,
+        undefined,
+        undefined,
+        Duration.milliseconds(10) // Very short timeout to trigger timeout quickly
+      );
+
+      // Verify that the result is a TestRunIdResult
+      expect(result).to.have.property('testRunId');
+      expect((result as TestRunIdResult).testRunId).to.equal(testRunId);
+
+      // Verify that the appropriate messages were logged
+      sinonAssert.calledWith(
+        debugStub,
+        sinon.match.string.and(sinon.match(testRunId))
+      );
+
+      const username = mockConnection.getUsername();
+      sinonAssert.calledWith(
+        loggerStub,
+        nls.localize('runTestReportCommand', [testRunId, username])
+      );
+    });
+  });
+
+  describe('run Apex and Flow tests in the same test run', () => {
+    let asyncTests: AsyncTests;
+
+    beforeEach(() => {
+      asyncTests = new AsyncTests(mockConnection);
+    });
+    it('should query the correct test results objects depending on the category', async () => {
+      const testQueueResult = {
+        done: true,
+        totalSize: 3,
+        records: [
+          {
+            Id: 'apex1',
+            ApexClassId: '01p000000000001',
+            Status: ApexTestQueueItemStatus.Completed,
+            TestRunResultId: 'result1'
+          } as ApexTestQueueItemRecord,
+          {
+            Id: 'flow1',
+            ApexClassId: null,
+            Status: ApexTestQueueItemStatus.Completed,
+            TestRunResultId: 'result2'
+          } as ApexTestQueueItemRecord,
+          {
+            Id: 'apex2',
+            ApexClassId: '01p000000000002',
+            Status: ApexTestQueueItemStatus.Completed,
+            TestRunResultId: 'result3'
+          } as ApexTestQueueItemRecord
+        ]
+      };
+
+      // Mock responses for Apex tests
+      const mockApexResults = {
+        done: true,
+        totalSize: 2,
+        records: [
+          {
+            Id: 'apexResult1',
+            QueueItemId: 'apex1',
+            StackTrace: null as string | null,
+            Message: null as string | null,
+            RunTime: 100,
+            TestTimestamp: '2023-01-01T10:00:00.000Z',
+            AsyncApexJobId: 'job1',
+            MethodName: 'testMethod1',
+            Outcome: 'Pass',
+            ApexLogId: null as string | null,
+            ApexClass: {
+              Id: '01p000000000001',
+              Name: 'TestClass1',
+              NamespacePrefix: null as string | null
+            }
+          },
+          {
+            Id: 'apexResult2',
+            QueueItemId: 'apex2',
+            StackTrace: null as string | null,
+            Message: null as string | null,
+            RunTime: 150,
+            TestTimestamp: '2023-01-01T10:00:00.000Z',
+            AsyncApexJobId: 'job1',
+            MethodName: 'testMethod2',
+            Outcome: 'Pass',
+            ApexLogId: null as string | null,
+            ApexClass: {
+              Id: '01p000000000002',
+              Name: 'TestClass2',
+              NamespacePrefix: null as string | null
+            }
+          }
+        ]
+      };
+
+      // Mock responses for Flow tests
+      const mockFlowResults = {
+        done: true,
+        totalSize: 1,
+        records: [
+          {
+            Id: 'flowResult1',
+            ApexTestQueueItemId: 'flow1',
+            Result: 'Pass',
+            TestStartDateTime: '2023-01-01T10:00:00.000Z',
+            TestEndDateTime: '2023-01-01T10:00:05.000Z',
+            FlowTest: { DeveloperName: 'TestFlow' },
+            FlowDefinition: {
+              DeveloperName: 'MyFlow',
+              NamespacePrefix: null as string | null
+            }
+          }
+        ]
+      };
+
+      // Setup mock to return different results based on query type
+      const mockToolingQuery = sandboxStub.stub(
+        mockConnection.tooling,
+        'query'
+      );
+      mockToolingQuery
+        .withArgs(sinon.match(/ApexTestResult/))
+        .resolves(mockApexResults)
+        .withArgs(sinon.match(/FlowTestResult/))
+        .resolves(mockFlowResults);
+
+      // Execute the test
+      const results = await asyncTests.getAsyncTestResults(testQueueResult);
+
+      // Verify that both queries were executed
+      expect(mockToolingQuery.calledTwice).to.be.true;
+
+      // Verify ApexTestResult query was called correctly
+      const apexQuery = mockToolingQuery
+        .getCalls()
+        .find((call) => call.args[0].includes('ApexTestResult'));
+      expect(apexQuery).to.exist;
+      expect(apexQuery.args[0]).to.include('FROM ApexTestResult');
+      expect(apexQuery.args[0]).to.include('WHERE QueueItemId IN');
+      expect(apexQuery.args[0]).to.include("'apex1'");
+      expect(apexQuery.args[0]).to.include("'apex2'");
+      expect(apexQuery.args[0]).to.not.include("'flow1'"); // Flow test should not be in Apex query
+
+      // Verify FlowTestResult query was called correctly
+      const flowQuery = mockToolingQuery
+        .getCalls()
+        .find((call) => call.args[0].includes('FlowTestResult'));
+      expect(flowQuery).to.exist;
+      expect(flowQuery.args[0]).to.include('FROM FlowTestResult');
+      expect(flowQuery.args[0]).to.include('WHERE ApexTestQueueItemId IN');
+      expect(flowQuery.args[0]).to.include("'flow1'");
+      expect(flowQuery.args[0]).to.not.include("'apex1'"); // Apex tests should not be in Flow query
+      expect(flowQuery.args[0]).to.not.include("'apex2'"); // Apex tests should not be in Flow query
+
+      // Verify results structure and categories
+      expect(results).to.have.length(2);
+
+      const apexResult = results.find(
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        (r: any) => r.category === TestCategory.Apex
+      );
+      expect(apexResult).to.exist;
+      expect(apexResult.records).to.have.length(2);
+
+      const flowResult = results.find(
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        (r: any) => r.category === TestCategory.Flow
+      );
+      expect(flowResult).to.exist;
+      expect(flowResult.records).to.have.length(1);
     });
   });
 });
