@@ -34,9 +34,9 @@ describe('Query Namespaces', async () => {
         username: testData.username
       })
     });
-    sandboxStub.stub(mockConnection, 'instanceUrl').get(() => {
-      return 'https://na139.salesforce.com';
-    });
+    sandboxStub
+      .stub(mockConnection, 'instanceUrl')
+      .get(() => 'https://na139.salesforce.com');
   });
 
   afterEach(() => {
